+++ src/main/java/dev/orne/i18n/I18nFixedString.java
@@ -26,6 +26,9 @@ import java.lang.ref.WeakReference;
 import java.util.Locale;
 import java.util.WeakHashMap;
 
+import javax.xml.bind.annotation.adapters.XmlAdapter;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
@@ -43,6 +46,7 @@ import jakarta.validation.constraints.NotNull;
  * @version 1.0, 2021-01
  * @since 0.1
  */
+@XmlJavaTypeAdapter(I18nFixedString.JaxbAdapter.class)
 public final class I18nFixedString
 implements I18nString {
 
@@ -181,5 +185,54 @@ implements I18nString {
     @Override
     public String toString() {
         return this.text;
+    }
+
+    /**
+     * JAXB adapter for {@code I18nFixedString} instances.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nFixedString
+     * @since 0.1
+     */
+    public static class JaxbAdapter
+    extends XmlAdapter<XmlI18nString, I18nFixedString> {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nFixedString value) {
+            return I18nString.JaxbAdapter.toXml(value);
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public I18nFixedString unmarshal(final XmlI18nString value) {
+            return I18nFixedString.from(I18nString.JaxbAdapter.fromXml(value));
+        }
+    }
+
+    /**
+     * JAXB adapter for {@code I18nFixedString} instances that marshalls
+     * all available translations.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nFixedString
+     * @since 0.1
+     */
+    public static class FullJaxbAdapter
+    extends JaxbAdapter {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nFixedString value) {
+            return I18nString.FullJaxbAdapter.toXml(value);
+        }
     }
 }