+++ src/main/java/dev/orne/i18n/I18nResourcesString.java
@@ -29,6 +29,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
 
+import javax.xml.bind.annotation.adapters.XmlAdapter;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
@@ -46,6 +49,7 @@ import jakarta.validation.constraints.NotNull;
  * @since 0.1
  */
 @JsonDeserialize(using=JsonDeserializer.None.class)
+@XmlJavaTypeAdapter(I18nResourcesString.JaxbAdapter.class)
 public class I18nResourcesString
 implements I18nString {
 
@@ -381,6 +385,56 @@ implements I18nString {
                     this.defaultText,
                     codes.toArray(new String[0]),
                     arguments.toArray(new Serializable[0]));
+        }
+    }
+
+    /**
+     * JAXB adapter for {@code I18nResourcesString} instances.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nResourcesString
+     * @since 0.1
+     */
+    public static class JaxbAdapter
+    extends XmlAdapter<XmlI18nString, I18nResourcesString> {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nResourcesString value) {
+            return I18nString.JaxbAdapter.toXml(value);
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public I18nResourcesString unmarshal(final XmlI18nString value) {
+            throw new UnsupportedOperationException(
+                    "Unmarshalling I18nResourcesString from XmlI18NString is not supported");
+        }
+    }
+
+    /**
+     * JAXB adapter for {@code I18nResourcesString} instances that marshalls
+     * all available translations.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nResourcesString
+     * @since 0.1
+     */
+    public static class FullJaxbAdapter
+    extends JaxbAdapter {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nResourcesString value) {
+            return I18nString.FullJaxbAdapter.toXml(value);
         }
     }
 }