+++ src/main/java/dev/orne/i18n/I18nStringMap.java
@@ -26,6 +26,9 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
+import javax.xml.bind.annotation.adapters.XmlAdapter;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
 import org.apache.commons.lang3.Validate;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
@@ -42,6 +45,7 @@ import jakarta.validation.constraints.NotNull;
  * @see I18nString
  * @since 0.1
  */
+@XmlJavaTypeAdapter(I18nStringMap.JaxbAdapter.class)
 public class I18nStringMap
 implements I18nString {
 
@@ -288,5 +292,54 @@ implements I18nString {
     @Override
     public String toString() {
         return this.getDefaultText();
+    }
+
+    /**
+     * JAXB adapter for {@code I18nStringMap} instances.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nStringMap
+     * @since 0.1
+     */
+    public static class JaxbAdapter
+    extends XmlAdapter<XmlI18nString, I18nStringMap> {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nStringMap value) {
+            return I18nString.JaxbAdapter.toXml(value);
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public I18nStringMap unmarshal(final XmlI18nString value) {
+            return new I18nStringMap(I18nString.JaxbAdapter.fromXml(value));
+        }
+    }
+
+    /**
+     * JAXB adapter for {@code I18nStringMap} instances that marshalls
+     * all available translations.
+     * 
+     * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+     * @version 1.0, 2021-02
+     * @see I18nStringMap
+     * @since 0.1
+     */
+    public static class FullJaxbAdapter
+    extends JaxbAdapter {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public XmlI18nString marshal(final I18nStringMap value) {
+            return I18nString.FullJaxbAdapter.toXml(value);
+        }
     }
 }