+++ src/test/java/dev/orne/i18n/I18nStringAsObjectContainer.java
@@ -24,29 +24,29 @@ package dev.orne.i18n;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
 /**
- * {@code I18nString} JSON/XML serialization test container.
+ * {@code I18nString} Jackson serialization test container.
  * 
  * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
- * @version 1.0, 2021-01
+ * @version 1.0, 2021-02
  * @since 0.1
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlRootElement(namespace=I18nStringContainer.TEST_NS, name=I18nStringContainer.ROOT_ELEMENT)
-public class I18nStringContainer {
+public class I18nStringAsObjectContainer {
 
-    public static final String TEST_NS = "http://orne.dev/i18n/test";
-    public static final String ROOT_ELEMENT = "container";
-    public static final String BEAN_ELEMENT = "bean";
-    public static final String BEAN_PROPERTY = "bean";
-
-    @XmlElement(namespace=TEST_NS, name=BEAN_ELEMENT)
+    @JsonFormat(shape = JsonFormat.Shape.OBJECT)
+    @XmlJavaTypeAdapter(I18nString.FullJaxbAdapter.class)
+    @XmlElement(namespace=I18nStringContainer.TEST_NS, name=I18nStringContainer.BEAN_ELEMENT)
     private I18nString bean;
 
     public I18nString getBean() {
@@ -69,7 +69,7 @@ public class I18nStringContainer {
         if (obj == null) { return false; }
         if (obj == this) { return true; }
         if (!getClass().equals(obj.getClass())) { return false; }
-        final I18nStringContainer other = (I18nStringContainer) obj;
+        final I18nStringAsObjectContainer other = (I18nStringAsObjectContainer) obj;
         return new EqualsBuilder()
                 .append(this.bean, other.bean)
                 .isEquals();