+++ src/test/java/dev/orne/i18n/jackson/I18nStringJacksonSerializationTest.java
@@ -38,11 +38,16 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import dev.orne.i18n.I18N;
+import dev.orne.i18n.I18nBilingualString;
 import dev.orne.i18n.I18nFixedString;
 import dev.orne.i18n.I18nResourcesString;
 import dev.orne.i18n.I18nString;
+import dev.orne.i18n.I18nStringAsObjectContainer;
+import dev.orne.i18n.I18nStringContainer;
 import dev.orne.i18n.I18nStringJacksonSerializer;
 import dev.orne.i18n.I18nStringMap;
+import dev.orne.i18n.I18nStringMapAsObjectContainer;
+import dev.orne.i18n.I18nStringMapContainer;
 import jakarta.validation.constraints.NotNull;
 
 /**